Replace VisualContainer by two components: for display and for editing
